+++ packages/api/src/test-utils/prisma-client.mock.ts
@@ -1,0 +1,10 @@
+export class PrismaClient {
+  [key: string]: any;
+  async $connect() {}
+  $on() {}
+  async $transaction<T>(cb: (tx: any) => Promise<T>): Promise<T> {
+    return cb({});
+  }
+}
+
+