+++ packages/webapp/src/app/providers/router.tsx
@@ -1,4 +1,6 @@
 import { ReactNode } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 
-export const RouterProvider = ({ children }: { children: ReactNode }) => <BrowserRouter>{children}</BrowserRouter>;
+export const RouterProvider = ({ children }: { children: ReactNode }) => (
+  <BrowserRouter future={{ v7_startTransition: true, v7_relativeSplatPath: true }}>{children}</BrowserRouter>
+);